+++ src/Configuration.jl
@@ -15,7 +15,7 @@ const TEST  = "test"
 haskey(ENV, "SEARCHLIGHT_ENV") || (ENV["SEARCHLIGHT_ENV"] = DEV)
 
 # defaults
-const SEARCHLIGHT_VERSION = v"0.12.0"
+const SEARCHLIGHT_VERSION = v"0.12.1"
 
 
 """
@@ -79,21 +79,26 @@ Dict{Any,Any} with 6 entries:
 ```
 """
 function read_db_connection_data(db_settings_file::String) :: Dict{String,Any}
-  db_conn_data =  if endswith(db_settings_file, ".yml")
-                    YAML.load(open(db_settings_file))
-                  elseif endswith(db_settings_file, ".jl")
-                    include(db_settings_file)
-                  end
+  db_conn_data::Dict =  if endswith(db_settings_file, ".yml")
+                          open(db_settings_file) do io
+                            YAML.load(io)
+                          end
+                        elseif endswith(db_settings_file, ".jl")
+                          include(db_settings_file)
+                        end
 
-  if  haskey(db_conn_data, "env") &&
-      db_conn_data["env"] != nothing
-    SearchLight.config.app_env = ENV["SEARCHLIGHT_ENV"] = db_conn_data["env"]
+  if  haskey(db_conn_data, "env") && db_conn_data["env"] != nothing
+    ENV["SEARCHLIGHT_ENV"] =  if db_conn_data["env"] == """ENV["GENIE_ENV"]"""
+                                ENV["GENIE_ENV"]
+                              else
+                                db_conn_data["env"]
+                              end
+
+    SearchLight.config.app_env = ENV["SEARCHLIGHT_ENV"]
   end
 
-  if  haskey(db_conn_data, SearchLight.config.app_env) &&
-      haskey(db_conn_data[SearchLight.config.app_env], "config") &&
-      db_conn_data[SearchLight.config.app_env]["config"] != nothing &&
-      isa(db_conn_data[SearchLight.config.app_env]["config"], Dict)
+  if  haskey(db_conn_data, SearchLight.config.app_env) && haskey(db_conn_data[SearchLight.config.app_env], "config") &&
+      db_conn_data[SearchLight.config.app_env]["config"] != nothing && isa(db_conn_data[SearchLight.config.app_env]["config"], Dict)
     for (k, v) in db_conn_data[SearchLight.config.app_env]["config"]
       setfield!(SearchLight.config, Symbol(k), ((isa(v, String) && startswith(v, ":")) ? Symbol(v[2:end]) : v) )
     end
@@ -102,7 +107,8 @@ function read_db_connection_data(db_settings_file::String) :: Dict{String,Any}
   return  if haskey(db_conn_data, SearchLight.config.app_env)
             db_conn_data[SearchLight.config.app_env]
           else
-            push!(SearchLight.SEARCHLIGHT_LOG_QUEUE, ("DB configuration for $(SearchLight.config.app_env) not found", :debug))
+            @error "DB configuration for $(SearchLight.config.app_env) not found"
+
             Dict{String,Any}()
           end
 end